ui: screen favorites, clean up folder, components, and layout
