+++ src/components/molecules/Header.tsx
@@ -1,6 +1,6 @@
 import {Image, StyleSheet, Text, TouchableOpacity, View} from 'react-native';
-import React from 'react';
 import {ArrowLeft, Bell, ChevronDown} from 'lucide-react-native';
+import React from 'react';
 import {colors} from '../../constants';
 
 type TPropsHeader = {
@@ -43,21 +43,28 @@ const Header: React.FC<TPropsHeader> = ({
 export default React.memo(Header);
 
 const styles = StyleSheet.create({
-  language: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    color: colors.primary,
-  },
   header: {
     flexDirection: 'row',
     paddingVertical: 24,
     paddingHorizontal: 16,
+    backgroundColor: colors.white,
     justifyContent: 'space-between',
     alignItems: 'center',
+    shadowColor: colors.disabled,
+    shadowOffset: {width: 0, height: 1},
+    shadowOpacity: 0.1,
+    shadowRadius: 2,
+    // Android elevation property
+    elevation: 5,
   },
   textLanguage: {
     fontSize: 18,
     fontWeight: '700',
+    color: colors.primary,
+  },
+  language: {
+    flexDirection: 'row',
+    alignItems: 'center',
     color: colors.primary,
   },
   logo: {
