+++ src/screens/Authentication/SignUpWithPhoneScreen.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import SignUpWithPhoneOrganism from '../../components/organisms/SignUpWithPhone/index.screen';
+
+export default function SignUpWithPhoneNumber() {
+  return <SignUpWithPhoneOrganism />;
+}
