+++ src/screens/Home/ProfileScreen.tsx
@@ -1,46 +1,8 @@
 import React from 'react';
-import {View, Text, Image, StyleSheet} from 'react-native';
+import ProfileOrganism from '../../components/organisms/Profiles/index.screen';
 
 const ProfileScreen = () => {
-  return (
-    <View style={styles.container}>
-      <Image
-        source={{uri: 'https://via.placeholder.com/150'}}
-        style={styles.profileImage}
-      />
-      <Text style={styles.name}>John Doe</Text>
-      <Text style={styles.bio}>
-        Software Developer at XYZ Company. Passionate about technology and
-        open-source projects.
-      </Text>
-    </View>
-  );
+  return <ProfileOrganism />;
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 20,
-    backgroundColor: '#f0f0f0',
-  },
-  profileImage: {
-    width: 150,
-    height: 150,
-    borderRadius: 75,
-    marginBottom: 20,
-  },
-  name: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginBottom: 10,
-  },
-  bio: {
-    fontSize: 16,
-    textAlign: 'center',
-    color: '#666',
-  },
-});
 
 export default ProfileScreen;
