+++ src/screens/IntroScreen.tsx
@@ -1,24 +1,8 @@
 import React from 'react';
-import CustomIntroSlider from '../components/organisms/CustomIntroSlider';
+import IntroductionOrganism from '../components/organisms/Introduction/index.screen';
 
-const slides = [
-  {
-    key: 'one',
-    title: 'Dive into a World of Flavor!',
-    image: require('../assets/images/intro-background.png'),
-  },
-  {
-    key: 'two',
-    title: 'Experience the Greatest!',
-    image: require('../assets/images/intro-background.png'),
-  },
-  {
-    key: 'three',
-    title: 'Savor the Burgers!',
-    image: require('../assets/images/intro-background.png'),
-  },
-];
+const IntroScreen = () => {
+  return <IntroductionOrganism />;
+};
 
-export default function IntroScreen() {
-  return <CustomIntroSlider slides={slides} />;
-}
+export default IntroScreen;
